requirements.txt
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
rich>=13.7.0
lxml>=5.1.0
pdfplumber>=0.11.0
streamlit>=1.35.0
notion-client>=2.2.1,<3.0.0